refactor: split runner kernel ebpf facades
